Move ObserverTrigger to live beside On
